Added User screen and updated Rps screen for lobby
Users can now get to a User screen after logging in or deciding to play
as guest. There is now a button to Play Game with leads to the Rps
screen which now displays the current game lobby. Users can now ready up
and lobby leaders can make others the leader and start the game.